+++ raphtory-graphql/src/main.rs
@@ -10,7 +10,6 @@ mod server;
 #[tokio::main]
 async fn main() {
     let graph_directory = env::var("GRAPH_DIRECTORY").unwrap_or("/tmp/graphs".to_string());
-    let vector_directory = env::var("VECTOR_DIRECTORY").unwrap_or("/tmp/vectors".to_string());
     RaphtoryServer::from_directory(&graph_directory)
         .run()
         .await
